transports: introduce `poll` callback
While streams now support polling for data, the transports making use of
these streams don't yet. Implement a `poll` callback with the same
reasoning as for the streams and provide a fallback of signalling
available data immediately if no polling function has been implemented
by the backend to not break existing backends.

Note that the WinHTTP backend doesn't have a poll implementation right
now. This is not by design, but simply because the author of this commit
doesn't have any idea on how to implement it.