transports: smart: implement receive keepalive
When receiving the packfile from a remote, git.git supports keepalive
packets to not cause the connection to disconnect if the remote takes
too long to start sending any bytes. Keepalive can be configured via
"receive.keepAlive", which is set to 5 seconds by default. If this
option is set to a positive value, then git will send a keepalive
packet, which is simply an empty pkt-line "0005\1".

Implement the same option for the smart protocol implementation.