regex: make sure to link against PCRE2's regex functions
PCRE2's header defines just the basic names but that means that we'll link
against libc's version while using PCRE2's struct definitions, leading to
crashes. To work around this, package maintainers have added prefixes to the
exported functions so we can link against them.

But the definitions in the header are still of the basic names so we need to
tell the compiler/linker to replace these names when linking.