transports: http: fix memory leak when unable to reconnect
On reading from an HTTP stream, when the server has closed a keepalive
socket or when receiving a 401 Unauthorized error, we try to re-connect
the HTTP stream in order to replay the authentication. When failing to
reconnect, we exit directly from the function and thus fail to cleanup
the request buffer.

Fix the memory leak by using `goto done` instead.