clone: detect out of memory when creating reflog message
When creating the reflog message "clone: from %s", we do not check the
call to `git_buf_printf` for an error. As we use `git_buf_cstr`
immediately afterwards, we may thus try to use the `git_buf_oom` buffer
without noticing. While this is unproblematic due to the buffer
containing a single NUL byte, only, we should certainly notice
out-of-memory situations early.

Fix this by checking the return value of `git_buf_printf`.