transports: ssh: set error code if username allocation fails
When trying to allocate the username that is required to ask for SSH
authentification methods, we check the return value of `git__strdup` but
will not set up a proper error code. Because of this,
`_git_ssh_setup_conn` may return successfully while there was in fact an
out of memory error.

Fix the issue by calling `git_error_set_oom` and setting the error code
to `-1`.