valgrind: be more lenient when suppressing libssh2 handshake leaks
The libssh2 library may leak memory when performing the session
handshake in various edge cases, like e.g. timeouts. As there is
multiple memleaks that may happen, depending on when the exact failure
happens inside of libssh2_session_handshake, we shouldn't specify a
detailed stack of which memleaks to suppress, but just suppress all that
happen with `libssh2_session_handshake`.