map: move `git_map` struct into POSIX layer
Our `git_map` interface is supposed to be an as-close-as-possible
emulation of the POSIX `mmap` function. Previously, though, we have been
using and abusing the structure to do all sorts of things when it comes
to keeping track of a pointer/length pair, completely ignoring that this
interface was intended as an `mmap` interface only. Because of this, we
have been refactoring code to instead use a new interface `git_mem`,
which is a higher-level abstraction of handling memory from various
sources.

As existing usage of `git_map` code has been replaced where possible, we
can now safely move it completely into the POSIX emulation layer. This
reduces its visibility across the codebase and properly highlights that
it is intended as a `mmap` replacement, only.