azure: tests: do not verify SSH host keys
The test suite online::clone::env will by default refuse any
certificates presented by the remote configured via environment
variables. In order for the tests to succeed with our instance of sshd,
we should either make the expected host key available via the
environment or alternatively just override it by setting
GITTEST_REMOTE_SSL_NOVERIFY.

This commit goes the easier way of just overriding the check, as the
point of the whole series is to first get the online::clone test suite
into a state where it's much easier to improve it at a later point in
time.