cmake: remove CheckPrototypeDefiniton module
The CMake module CheckPrototypeDefiniton has been introduced in CMake
v3.0, which is why we had to include a fallback copy of it in our own
modules in case somebody was compiling libgit2 with an older version of
CMake back when the minimum version was v2.8.11. As we have since raised
the minimum version to v3.5.1, we can now remove the module from our own
modules and simply always use the one provided by upstream.