cmake: manage project version in build instructions
Currently, the project version is hidden inside of our "version.h"
header file and CMake extracts required information out of that file via
a regular expression. This is unwieldy and completely backwards, as it
should be the CMake file that sets the project's version and then puts
it into a configured file.

Move version information into the call to `PROJECT` and use
`CONFIGURE_FILE` to configure the "version.h.in" header. The configured
header will now be configured into a "include/git2/" directory together
with the "features.h" header and then finally gets installed into the
target directory. This also removes the awful hack of directly adding
our binary directory into the include paths, as there is now a
subdirectory "include" in our binary directory for configured headers.