cmake: always set minimum required version first
Technically speaking, it is necessary to always set the minimum required
CMake version first before calling any other function in a CMake
project. Thus we should move the call to `CMAKE_MINIMUM_REQUIRED` before
the call to `PROJECT`, especially so as `PROJECT` may use features of
CMake which have been introduced with a certain version, only.