cmake: introduce new LinkDependency function
The libgit2 library has quite a lot of different dependencies that may
or may not be optional. Currently, the dependency handling code is a
_huge_ mess which nobody is able to grasp at all. Most of it is
currently done in the CMake build instructions of the source directory,
but they are split up to multiple locations, use global tracking via
weird variables and generally have been a source for confusion.

Modern CMake actually helps to sort out this mess by using targets all
the way through. Instead of manually tracking include directories, link
directories and flags, we can just set up a target for each dependency
that is completely self contained and has all information required to
link against it.

Introduce a new helper function `LINK_DEPENDENCY` that aims to help
migrating towards that target. The assumption is quite simple: for each
dependency, there will be a "cmake/DependencyFoo.cmake" file that gets
included by the helper. This dependency file is expected to _always_
export a target, even for dependencies which are not going to be used.
Thus, in the case where a dependency is found, the module should export
a normal target that contains linking and preprocessor information. In
the case where no dependency is found or in the case where it is not to
be used, the module is expected to export an empty target that simply
does nothing. Like this, every submodule can be completely
self-contained and we do not need to treat them specially in the main
build module.