repo: consolidate low-level "layout" details in a separate header
As the number of places that expect a actual .git directory to exist increases,
users of custom repository have started to encounter a varying amount of
"brokenness", depending on how custom their repository are.

As a first step to alleviate this, all used "names" are gathered in a new
header, the `git_repository_layout` struct consolidates the "file-system
resolution logic" already in place, and all users of direct file-system access
are changed to go through that system.