creds: fix missing include of <libssh2.h>
The cred.h header adds forwards declarations for libssh2 structures if
libssh2 is not being used. The decision is based on whether or not
LIBSSH2_VERSION is defined, which is in turn declared by <libssh2.h>.
But as we never include this header, it is a matter of luck whether
anybody else has included it transitively already.

Fix the issue by including <libssh2.h> if GIT_SSH is defined.