test: add an azure repos test
We currently talk to Azure Repos for executing an online test
(online::clone::path_whitespace).  Add a simpler test to talk to Azure
Repos to make it obvious that strange test failures are not likely the
whitespace in the path, but actually a function of talking to Azure
Repos itself.