changelog: add more newly-accepted urls
I encountered some problematic URLs,
and was delighted to see that they were already fixed.
I figured I may as well add them to the changelog.

For the record, URLs with no path used to be rejected.
That is arguably correct, but command line git accepts them.

URLs with a path of / and a non-standard port
used to have their port completely ignored!